requirements.txt
et-xmlfile==1.0.1
future==0.18.2
jdcal==1.4.1
openpyxl==3.0.5
python-whois==0.7.3
